oasp/oasp4j#358: reworked module structure as suggested by Simon, now startes only contain spring.factories, results however in a magnitude of modules
