+++ bom/pom.xml
@@ -316,12 +316,62 @@
       </dependency>
       <dependency>
         <groupId>io.oasp.java.modules</groupId>
+        <artifactId>oasp4j-cxf-client</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.oasp.java.modules</groupId>
+        <artifactId>oasp4j-cxf-client-rest</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.oasp.java.modules</groupId>
+        <artifactId>oasp4j-cxf-client-ws</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.oasp.java.modules</groupId>
+        <artifactId>oasp4j-cxf-server</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.oasp.java.modules</groupId>
+        <artifactId>oasp4j-cxf-server-rest</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.oasp.java.modules</groupId>
+        <artifactId>oasp4j-cxf-server-ws</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.oasp.java.modules</groupId>
         <artifactId>oasp4j-starter-cxf-client</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
         <groupId>io.oasp.java.modules</groupId>
+        <artifactId>oasp4j-starter-cxf-client-rest</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.oasp.java.modules</groupId>
+        <artifactId>oasp4j-starter-cxf-client-ws</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.oasp.java.modules</groupId>
         <artifactId>oasp4j-starter-cxf-server</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.oasp.java.modules</groupId>
+        <artifactId>oasp4j-starter-cxf-server-rest</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.oasp.java.modules</groupId>
+        <artifactId>oasp4j-starter-cxf-server-ws</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
