+++ modules/cxf-client-rest/pom.xml
@@ -8,16 +8,24 @@
     <version>dev-SNAPSHOT</version>
   </parent>
   <groupId>io.oasp.java.modules</groupId>
-  <artifactId>oasp4j-starter-cxf-server</artifactId>
+  <artifactId>oasp4j-cxf-client-rest</artifactId>
   <version>${oasp4j.version}</version>
   <packaging>jar</packaging>
   <name>${project.artifactId}</name>
-  <description>Support for service server based on Apache CXF.</description>
+  <description>Support for consuming REST services as client based on Apache CXF.</description>
 
   <dependencies>
     <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>oasp4j-cxf-server</artifactId>
+      <artifactId>oasp4j-cxf-client</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>oasp4j-json</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.cxf</groupId>
+      <artifactId>cxf-rt-rs-client</artifactId>
     </dependency>
     <dependency>
       <groupId>io.oasp.java.modules</groupId>
