+++ modules/cxf-client/src/main/java/io/oasp/module/cxf/common/impl/client/CxfClientAutoConfiguration.java
