+++ modules/cxf-server-rest/src/main/java/io/oasp/module/cxf/common/impl/server/rest/CxfServerRestAutoConfiguration.java
