+++ modules/cxf-server-ws/pom.xml
@@ -8,16 +8,24 @@
     <version>dev-SNAPSHOT</version>
   </parent>
   <groupId>io.oasp.java.modules</groupId>
-  <artifactId>oasp4j-starter-cxf-server</artifactId>
+  <artifactId>oasp4j-cxf-server-ws</artifactId>
   <version>${oasp4j.version}</version>
   <packaging>jar</packaging>
   <name>${project.artifactId}</name>
-  <description>Support for service server based on Apache CXF.</description>
+  <description>Support for providing web-services (SOAP) as server based on Apache CXF.</description>
 
   <dependencies>
     <dependency>
       <groupId>${project.groupId}</groupId>
       <artifactId>oasp4j-cxf-server</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.cxf</groupId>
+      <artifactId>cxf-rt-frontend-jaxws</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-web-services</artifactId>
     </dependency>
     <dependency>
       <groupId>io.oasp.java.modules</groupId>
