+++ modules/starter-cxf-server-rest/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=io.oasp.module.cxf.common.impl.server.rest.CxfServerRestAutoConfiguration
