+++ modules/starter-cxf-server/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-io.oasp.module.cxf.common.impl.server.CxfServerAutoConfiguration,\
-io.oasp.module.cxf.common.impl.server.rest.CxfServerRestAutoConfiguration,\
-io.oasp.module.cxf.common.impl.server.soap.CxfServerSoapAutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=io.oasp.module.cxf.common.impl.server.CxfServerAutoConfiguration
